Fix small bug with window popup
git-svn-id: https://vcs.maemo.org/svn/omweather/trunk@51 5aa3d57c-301e-0410-a7f2-8fec96c11689

+++ omweather/src/weather-home_common.h
@@ -52,7 +52,7 @@ typedef struct
   gchar hmid[5];        /* Humidity */
   int wind_speed;       /* Wind Speed */
   int wind_gust;        /* Wind Gust */
-  int vis;              /* Visibilty */
+  float vis;              /* Visibilty */
   gchar wind_title[11]; /* Wind Title */
   gchar temp[4];        /* Temperature */
   time_t begin_time;    /* Begin time party of the day */
